+++ src/pytest_shared_session_scope/fixtures.py
@@ -173,12 +173,14 @@ def shared_session_scope_fixture(
                     )
                     tests_missing -= set(tests_run_in_worker)
                     is_last = not tests_missing
-                    if not is_last:
-                        metadata_storage.write(
-                            metadata_identifier,
-                            json.dumps(list(tests_missing)),
-                            fixture_values,
-                        )
+                    # TODO: The fixture that closes shouldn't need to write here
+                    # But there are issue with getfixturevalue.
+                    # if not is_last:
+                    metadata_storage.write(
+                        metadata_identifier,
+                        json.dumps(list(tests_missing)),
+                        fixture_values,
+                    )
 
                 if is_last:
                     _send_last(res, CleanupToken.LAST)